Ensure creation timestamp of PodSandbox is valid
The creation timestamp in the status of a sandbox was only set after
starting the sandbox container successfully. If there was any error, the
status was still added to the sandbox store to ensure resources will be
cleaned up. But the stored status has an empty creation timestamp in
this case, which was then converted to a negative number
-6795364578871345152 when ListPodSandbox was called.

This negative number breaks the contract of PodSandbox API:

```
// Creation timestamp of the sandbox in nanoseconds. Must be > 0.
int64 created_at = 4;
```

And it leads to weird output when viewing the sandboxes via crictl:

```
$ crictl pods
POD ID              CREATED             STATE               ...
819ce20e6bd8d       269 years ago       NotReady            ...
```

This commit sets the creation timestamp after creating the sandbox
container successfully, by which the timestamp is already available, to
ensure the field is always a valid value in the store. Besides, it sets
the WithoutRefreshedMetadata option when getting the information of the
container to avoid an unnecessary lookup as the creation timestamp is
already available in the metadata returned by the container creation
function.

Signed-off-by: Quan Tian <qtian@vmware.com>